requirements.txt
django==4.0.2
djangorestframework==3.13.1
requests==2.27.1
django-cors-headers==3.11.0
pytest==7.0.1
python-decouple==3.6
django-filter==21.1
psycopg2==2.9.3